requirements/test.txt
git+https://github.com/edx/xblock-sdk.git@c07c867d5b3a5eb8a4a241b218083fb1d3416fc9#egg=xblock-sdk

coverage==3.7.1
ddt==0.8.0
django-nose==1.3
mock==1.0.1
pep8==1.5.7
pylint<1.0

